backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
reportlab==4.0.7
cairosvg==2.7.1
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
python-docx==1.1.0
docxtpl==0.20.2
asyncpg==0.29.0
python-barcode[images]==0.16.0
Pillow==10.1.0